+++ projects/NeRF-Det/nerfdet/__init__.py
@@ -1,0 +1,3 @@
+from .scannet_multiview_dataset import ScanNetMultiViewDataset
+
+__all__ = ['ScanNetMultiViewDataset']
